+++ src/lib/analytics/range-delta.ts
@@ -33,53 +33,27 @@ import {
   readBestGranularityRollups,
   type RollupBucketRow,
 } from "@/lib/rollups/measurement-read-wmy";
+import {
+  ANALYTICS_RANGES,
+  rangeWindowDays,
+  type AnalyticsRange,
+  type RangeDeltaResult,
+  type WindowAggregate,
+} from "@/lib/analytics/range-shared";
 
-const DAY_MS = 24 * 60 * 60 * 1000;
-
-/** The user-selectable ranges. */
-export const ANALYTICS_RANGES = ["7d", "30d", "90d", "1y"] as const;
-export type AnalyticsRange = (typeof ANALYTICS_RANGES)[number];
-
-/** Trailing-window length, in days, for each range. */
-const RANGE_WINDOW_DAYS: Record<AnalyticsRange, number> = {
-  "7d": 7,
-  "30d": 30,
-  "90d": 90,
-  "1y": 365,
+// The range constants + result shapes live in the client-safe
+// `range-shared` module so the insights client bundle never pulls the
+// server-only rollup readers (and `pg`) through a range import. Re-exported
+// here so existing server-side callers keep importing from `range-delta`.
+export {
+  ANALYTICS_RANGES,
+  rangeWindowDays,
+  type AnalyticsRange,
+  type RangeDeltaResult,
+  type WindowAggregate,
 };
 
-export function rangeWindowDays(range: AnalyticsRange): number {
-  return RANGE_WINDOW_DAYS[range];
-}
-
-/** A composed window aggregate. `mean`/etc. are null when the window is empty. */
-export interface WindowAggregate {
-  count: number;
-  min: number | null;
-  max: number | null;
-  mean: number | null;
-  sum: number | null;
-}
-
-export interface RangeDeltaResult {
-  range: AnalyticsRange;
-  windowDays: number;
-  /** The granularity the underlying rollup read resolved against. */
-  granularity: string;
-  current: WindowAggregate;
-  previous: WindowAggregate;
-  /**
-   * `current.mean - previous.mean`, or null when either window has no mean
-   * (no data in the current or prior window) — never a misleading 0.
-   */
-  delta: number | null;
-  /**
-   * `delta / previous.mean` as a fraction (0.03 = +3 %), or null when the
-   * prior window has no mean / a zero mean (no divide-by-zero, no misleading
-   * 0 %). The client renders "no prior-period data" in that case.
-   */
-  deltaPct: number | null;
-}
+const DAY_MS = 24 * 60 * 60 * 1000;
 
 /**
  * Slice the resolved buckets into the current half (`[now-N, now)`) and the
